requirements.txt
Flask==2.3.2
Pillow==10.0.0
tensorflow==2.13.0
opencv-python==4.7.0.72
numpy==1.25.0